Safari-609.1.11.1/LayoutTests/fast/dom/HTMLScriptElement/defer-onbeforeload-expected.txt

Checks that deferred scripts fire onbeforeload immediately and that it is cancellable.
cancelled onbeforeload defer
onbeforeload defer
onbeforeload external
external
defer
